requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
scikit-learn
matplotlib
seaborn
pandas
google-generativeai
python-multipart
Pillow
